requirements.txt
# Core dependencies
boto3>=1.34.0
aioboto3>=12.0.0
supabase==2.5.0
python-dotenv==1.0.0

# Document processing
PyMuPDF==1.24.5
pandas==2.2.2
openpyxl==3.1.2

# Web and API
streamlit==1.35.0
aiohttp==3.9.5
requests==2.32.3

# Utilities
pydantic==2.7.4
python-multipart==0.0.9
typing-extensions==4.12.2

# Development/Testing
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.1.0